+++ Online/Entity/OnlinePhysicalObject.cs
@@ -72,7 +72,7 @@ namespace RainMeadow
             beingMoved = false;
         }
 
-        protected override EntityState MakeState(ulong tick, OnlineResource resource)
+        protected override EntityState MakeState(uint tick, OnlineResource resource)
         {
             if (resource is WorldSession ws && !LobbyManager.lobby.gameMode.ShouldSyncObjectInWorld(ws, apo)) throw new InvalidOperationException("asked for world state, not synched");
             if (resource is RoomSession rs && !LobbyManager.lobby.gameMode.ShouldSyncObjectInRoom(rs, apo)) throw new InvalidOperationException("asked for room state, not synched");
